add restoration from backup option

+++ code/UI.py
@@ -4,6 +4,7 @@ action_options = """
 1. View movies list
 2. Add new movie
 3. Delete a movie
+4. Restore movies list from backup
 
 0. Quit
 """
@@ -14,9 +15,6 @@ def ui() -> bool:
     while True:
         try:
             user_input = int(input('Choose an option: '))
-            if user_input < 0 or user_input > 3:
-                print('Invalid input')
-                continue
             break
         except ValueError:
             print('I need a number')
@@ -32,19 +30,27 @@ def ui() -> bool:
                 print('No movies found')
         case 2:
             Film.add_by_api(input('Enter movie title: '))
+            Film.create_backup()
             print('Movie added')
         case 3:
-            for film in Film.all():
-                print(film.__repr__())
-            try:
-                user_input = int(input('Choose a movie: '))
-                if user_input not in [film.id for film in Film.all()]:
-                    print('Invalid input')
-                    return True
-                Film.get(user_input).delete()
-                print('Movie deleted')
-            except ValueError:
-                print('I need a number')
+            if Film.all():
+                for film in Film.all():
+                    print(film.__repr__())
+                try:
+                    user_input = int(input('Choose a movie: '))
+                    if user_input not in [film.id for film in Film.all()]:
+                        print('Invalid input')
+                        return True
+                    Film.get(user_input).delete()
+                    Film.create_backup()
+                    print('Movie deleted')
+                except ValueError:
+                    print('I need a number')
+            else:
+                print('No movies found')
+        case 4:
+            Film.load_from_backup()
+            print('Movies loaded')
         case _:
             print('I didn\'t get your input')
     return True
